server/requirements.txt
jq
flask 
flask-cors 
requests
python-dotenv 
langchain-google-genai 
langchain-community 
langchain-pinecone
pinecone-client
firebase
firebase-admin
twilio